admin_api/requirements.txt
fastapi
uvicorn
sqlite-utils
psycopg2-binary

